Start actually using async HTTP requests
I did all this work to use async libraries and whatnot, and then
set up my code to "await" each request, essentially running it
synchronously.

There's probably a better way to do this, but for now, I'm using
most of the codebase I have and just creating tasks.

This is definitely not complete.